Created app bar component
