+++ src/App.js
@@ -1,11 +1,14 @@
 import './App.css';
-import LandingSection from './landing-page/LandingSection';
-import ScheduleSection from './landing-page/ScheduleSection';
-import SponsorsSection from './landing-page/SponsorsSection';
+
+import AppBar from './components/AppBar';
+import LandingSection from './containers/landing-page/LandingSection';
+import ScheduleSection from './containers/landing-page/ScheduleSection';
+import SponsorsSection from './containers/landing-page/SponsorsSection';
 
 function App() {
   return (
     <div className="App">
+      <AppBar />
       <LandingSection />
       <ScheduleSection />
       <SponsorsSection />
