+++ src/containers/landing-page/LandingSection.js
@@ -1,0 +1,21 @@
+function LandingSection() {
+  return (
+    <section className="hero is-info is-fullheight">
+        <div className="hero-head">
+        </div>
+
+        <div className="hero-body">
+            <div className="container has-text-centered">
+            <p className="title">
+                Instituto de Computação da Unicamp
+            </p>
+            <p className="subtitle">
+                Unicamp de Portas Abertas 2021
+            </p>
+            </div>
+        </div>
+    </section>
+  );
+}
+
+export default LandingSection;
